Use a correct approach to fetch clip-rect
For default-sized group, its size is determined by its children's size.
And for a default sized node, their clip-rect is not the (content) rect
of its parent, but its ancestor nodes' clip-rect! To get the correct
clip-rect, we need to intersect current clip-rect with all ancestors
follow the node link to the root ancestor.

Closes #36