Fix image-button issue: partial image not rendered
Adjust DrawingContext.DrawImage overload from
  `DrawImage(ITexture image, Rect rectangle)`
to
  `DrawImage(ITexture image, Rect rectangle, Point uvMin, Point uvMax)`

Why?
Because texture coordinates and corresponding size are the actual key
points when rendering a texture with GPU, not the pixel offset, which
will finally be calculated into texture coordinates.

Add new style property `ObjectPosition` to control texture uv when
rendering an image, which has been applied to `ImageButton`.