Don't throw exception for Zero-sized clip-rect.
Case: a child clip rectangle which doesn't intersect with the parent
clip rectangle, then the pushed clip rectangle is `Rect.Zero`.

Allowing zero-rectangle aims to make the push/pop API work when pushed
rectangle is `Rect.Zero`. If we limit the push rectangle to be non-zero,
we need to check everywhere when trying to push a clip rectangle.
That will bring too many changes.