Fixed #57
The order of controls inside a CollapsingHeader, which is initially
closed, is inverted.

The key point: If a control API "A" call that have already created its
node, it will not update its position in parent's children list any
more. And when new control API "B" are called before "A", B will be add
to parent's children list at a position after "A". In this case "A" will
 be layouted after "B", while "A" control API is called before "B".